+++ python/GafferDispatchTest/__init__.py
@@ -61,6 +61,7 @@ from .StatsApplicationTest import StatsApplicationTest
 from .FileListTest import FileListTest
 from .DeleteFilesTest import DeleteFilesTest
 from .CopyFilesTest import CopyFilesTest
+from .RenameFilesTest import RenameFilesTest
 
 if __name__ == "__main__":
 	import unittest